requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.4
Pillow>=10.3.0
aiofiles>=23.2.1

# PyTorch CPU-only (nhẹ hơn GPU ~700MB vs ~2GB)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
